+++ CHANGELOG.md
@@ -13,6 +13,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Updated French path in French language
 - Added text to the feedback form to clearly indicate that the text area is required
 - Removed signup confirmation/validation pages from search engine indexing
+- Added projects disclaimer to the main body text of each page
+
 
 ## Fixed